Responding to improvement feedback.
Adds in error handling for Server requests. Added in timeouts
Included retry mechanism for failed http requests.

Added in a debug mode to route requests to a local server

Added Disable Device call to the SDK

Abstracted the bundle from intent when creating a new notification.

Generated a unique id from notifications so they do not overwrite the old notification.