Added webview status to have better granularity
Now the resize script will be called when webview starts loading and also when the page is finished loading, removing the possibility of hiding the webview for long time caused due to delayed resource loading.
For Dialog implementation to have same experience as previous implementation, the resize script will be processed only if the webview status is loaded.