Resize WebView
Use relative layout gravity instead to place the in-app in desired location.
Resizing logic indirectly moved to webview instead of its holding parent fragment.
TODO: Need to delete out the commented code and Log comments before merging it to master.
Kept it for direct reference for changes to come.
enum Layout is introduced for webviewClient to understand and animate accordingly. Changes will be added in next commits