[MOB-4068] - Reset jwtToken when logging out
For JWT enabled projects using Android SDK, a logout and login ends up disabling the user but not registering the user again(stopping them from getting push notification). Reason being the auth not reset because the authHandler kept returning the same authtoken and was not set again. Instead it should be set to null and then next time user logs in, the code can continue normally