[MOB-4998] - OfflineMode ConnectivityManager Improvements
Some improvements in ConnectivityManager.

1. By default, isConnected is set to true. to avoid situations of db bloat up and no listener registered. This will also make sure that an app restart will eventually reset the flag, in case where internet was available but callbacks were not registered.
2. ConnectivityManager also listens to onCapabilitiesChanged, which can indicate more granular detail of if internet is available or not. isConnected is set to true as soon as internet is available.
Currently, opposite case is not true as during testing, it has been seen as when one network out of the two gets disconnected, sometimes, false callbacks are made and another callback with internet available is not called.

Added a TODO statement which captures even larger context to make ConnectivityManager even better. Might have to store multiple network that device is connected to and as long as one exists isConnected switch can remain ON.