Taking account of multiple network
Device can be connected to multiple network at once. Mostly Wifi and cellular. Previous implementation would switch back to offline if either of the one got disconnected, thereby falsely picturing device being offline.
This implementation takes multiple network into consideration and saves them in a Set. Only if all the network are disconnected, only then `isConnected` is flipped to false.
onCapabilitiesChanged is another callback that can give even granular controls over if internet is actually present or not, but to keep it simple, this is a good improvement to existing functionality of offline mode.