Remove all other channels which are not active
Remove all the other channels which are not actively showing notification on the notification tray.
Warning: Have to check if it only removes the notification created by iterable package? or the entire app's other channels as well.
We might have to add iterable_ or some kind of identifier to delete only Iterable ones.