[MOB-5277] - Enforced Encryption
Implementation of enforced encryption with fallback to unencrypted shared preference in case the initialization of EncryptedSharedPreference throws exception.
Introduced enforcedEncryption Key in config. This will enable app to enforce encryption. This may crash the app due to some devices reported to not work well with Android Crypto library. Unless required, enforced encryption is not recommended.