Part 4 - Chaging deprecated method
Allowing online request processor to run the execution on executor, allowing for async approach.
Offline request processor is untouched and will follow the Tasks one by one as it might need to stop processing if one of them fails and avoid data loss.
This approach seems not harmful for online request processor