Mutable flag for buttons with remote input
This is an effort to make the test of button with action with remote inputs be mutable.
The tests are all passing. But I am also feeling, whether to mark pendingIntent as Mutable or Immutable can be decided based on if a mutable intent is required. Remote input might require mutable intent as the data from keyboard could change the intent. But may be not other buttons which do not necessarily open the app. But providing mutable doesnt cause big harm is what Im thinking.