[MOB-9211] - InApp resizing fix
Removing treeobserver as the html would keep calling this callback even after the page was finished loading. This method was used as an additional measure to make sure inapps are resized in situation when inapps were still not displayed after finished loading.
This fix adds delay of 100ms which seems to resolve inapps on Pixel 6pro and still making it all work on Pixel 8 device, needing the SDK to not rely on tree observers.