Instruction files
Firebase setup md explains structure of integration test project and also required configuraiton to make it work
Readme.md is integration test project's read indicating what all feature work is going to be included in the project. and eventually how to run those tests and requirements for those.
build.gradle is added as a boiler plate code. It is non-checked generated files. A revisit to remove unncessary code should be done as needed. Current aim to to achieve end-to-end integration test workflow setup. Hence not targetting to line by line check if something better is availabel OR if a line is really needed or not
google-service.json.template is added just for reference. can totally delete this file later.
proguard rules added by Cursor. Not sure if we need this. There is a open issue created by developer indicating we dont need it anymore. But need to verify that.
Super important part of this commit is the settings.gradle file which will now allow integration test to access the api and api-ui folders